recv_request_vote: Mirror request term in pre-vote response.
Based on the solution used in ScyllaDB [1] to prevent a cluster deadlock
when a node with a stale term reconnects to the cluster but its vote
is needed to form quorum. Using the stale term in the response would
lead to the other nodes ignoring the response and would result in an
unavailable cluster.

[1]: https://github.com/scylladb/scylla/blob/e5956fee8af8ad7b6ba1e797f269729406afd7ef/raft/fsm.cc#L794