raft: Add 'constructors' for struct raft, raft_fsm & raft_io
This allows the raft library to allocate enough space for the structs
without depending on the space allocated by the user that could be
compiled against an older version of libraft.

I eventually want to restructure the header files and hide internal
data structures that should not have been exposed.