uv_segment: Introduce a new disk format for segments.
The (open) segments now contain the first index of the encoded log
entries and a checksum over the format and first index fields.

The 24 byte header looks like:
  _________________________________________
 |[0 - 7] |      [8-15] | [16-19] | [20-23]|
 |FORMAT  | FIRST_INDEX |     CRC |  UNUSED|
 |________|_____________|_________|________|

This allows us to no longer guess the start index and do more
direct checks. This also allows removing the barrier when taking a
snapshot.