test_replication: turn the assert_config function into a macro
Using functions for assertions has the downside that when a test fails munit
will report a line number that points to the assert function, not to the test
function, making it more difficult to spot the exact point where the test fails.

When using macros, munit will report the appropriate offending line of the test
function (the one where the macro is invoked).

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>