Remove code duplication for caching the configuration of a snapshot
Every time we restore a snapshot we must also cache in r->configuration_previous
the configuration contained in the snapshot (which the last known committed
configuration at the time the snapshot was taken). This is necessary in case the
log gets truncated and loses track of the entry that originally contained that
configuration.

We restore a snapshot in two cases:

- when receiving a SnapshotInstall RPC
- when starting up

Instead of duplicating the caching logic in both places, move the caching logic
to the snapshotRestore() utility, which is responsibile for all bookkeeping
needed upon snapshot restore, no matter when and where it happens.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>