Make r->configuration_previous always contain snapshot configuration
After successfully taking and persisting a snapshot, we generally cache into
r->configuration_previous the configuration that was included in the snapshot.

Technically speaking we don't need to do that if between the time the snapshot
was started and the time it completed new configuration changes were
committed (those won't be truncated yet).

However, for consistency, change the code to populate r->configuration_previous
no matter what, so we preserve the invariant that r->configuration_previous
always contains a copy of the configuration included in the most recent
snapshot.

Also, make the code slightly less indirect by using configurationCopy() instead
of the configurationBackup() helper, which has now just this one call site and
isn't worth be kept (it will be removed in a follow-up commit).

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>