start: Remove redundant snapshot configuration caching logic
When we start up we first restore the last persisted snapshot (if any) and then
all persisted entries.

As part of resting the last persisted snapshot, we cache its associated
configuration in r->configuration_previous, for rollbacks.

We don't need to cache it again in restoreConfigurations(), which is called when
persisted entries are restored.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>