fixture: don't override user data in ioMethodAsyncWork
The `raft_io_async_work->data` attribute should not be set or read by a
`raft_io` implementation, since it's meant for user data.

The `ioMethodAsyncWork` was setting it to the `struct raft` object associated
with the fixture itself. Howver this was working only incidentally, because the
user code in `replication.c` is also setting it to the same `struct raft`
object.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>