Send AppendEntries result if term changes when installing a snapshot
It's okay to send an AppendEntries result to a leader that was not the same of
the one that originally sent the snapshot, or that has bumped its term.

This makes the logic a bit simpler, since we don't need to track what the term
was at the time the InstallSnapshot RPC was first received.

It also can speed up synchronization in case the new leader does not yet know
the progress of the follower.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>