election: Only consider persisted state when running elections
At the moment we correctly apply pending configurations only once they are
persisted. However, we don't do the same when running an election, where we use
not-yet-persisted log indexes, leading to incosistencies and violation of Raft
invariants.

This commits makes RequestVote messages include the index/term of the last
persisted entry, not the one in the in-memory log cache.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>