configure: Remove redundant with_lz4 checks
If $have_lz4 is set to "yes", then $with_lz4 must be different than "no" and
there's no need to explicitly check that.

That's because the only time that $have_lz4 can be "yes" is when $with_lz4 is
not "no", due to this AS_IF:

AS_IF([test "x$have_uv" = "xyes"],
      # libuv is used
      [AS_IF([test "x$with_lz4" != "xno"],
             [PKG_CHECK_MODULES(LZ4, [liblz4 >= 1.7.1], [have_lz4=yes], [have_lz4=no])],
             [have_lz4=no])

If $with_lz4 is "no", then $have_lz4 is guaranteed to be "no" as well.

Note that the various "have_xxx" variables are basically the "result" of
processing the associated "enable_xxx"/"with_xxx" variables (which map directly
to command line arguments). So once a "have_xxx" has been set, there should be
no need to inspect again the associated command line flags.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>