test_uv_append: run loop to avoid race.
In the append/noSpaceUponWrite test, here's a race condition where
the prepare step of an open segment is started while the disk is still
full, but only completes after the balloon file on disk is removed.
After removal of the balloon file it is assumed that the APPEND(5, 64)
step succeeds, but the append command is linked to the segment that
will fail to prepare due to a full disk, resulting in a test assertion
failure.

Running the eventloop a couple of times (hopefully) fixes this race.

Signed-off-by: Mathieu Borderé <mathieu.bordere@canonical.com>