client: Only send timeout now message if persisted state is up-to-date.
Receivers of the "timeout now" message ignore the message if their
in-memory logs are not up-to-date or if they are persisting log entries.
When the message is ignored, the sender will not retry it which can lead
to failed leadership transfers that could have been avoided. Instead,
the leader only sends the message when it knows the follower has
persisted all log entries, based on the follower's match_index.

Signed-off-by: Mathieu Borderé <mathieu.bordere@canonical.com>